+++ inc/wv-register-taxonomy.php
@@ -14,19 +14,20 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 $labels = array(
 	'name' => esc_html__( 'Videos Categories', '%TEXTDOMAIN%' ),
-	'singular_name' => esc_html__( 'Videos Type', '%TEXTDOMAIN%' ),
+	'singular_name' => esc_html__( 'Videos Category', '%TEXTDOMAIN%' ),
 	'search_items' => esc_html__( 'Search Videos Categories', '%TEXTDOMAIN%' ),
 	'popular_items' => esc_html__( 'Popular Videos Categories', '%TEXTDOMAIN%' ),
 	'all_items' => esc_html__( 'All Videos Categories', '%TEXTDOMAIN%' ),
-	'parent_item' => esc_html__( 'Parent Videos Type', '%TEXTDOMAIN%' ),
-	'parent_item_colon' => esc_html__( 'Parent Videos Type:', '%TEXTDOMAIN%' ),
-	'edit_item' => esc_html__( 'Edit Videos Type', '%TEXTDOMAIN%' ),
-	'update_item' => esc_html__( 'Update Videos Type', '%TEXTDOMAIN%' ),
-	'add_new_item' => esc_html__( 'Add New Videos Type', '%TEXTDOMAIN%' ),
-	'new_item_name' => esc_html__( 'New Videos Type', '%TEXTDOMAIN%' ),
+	'parent_item' => esc_html__( 'Parent Videos Category', '%TEXTDOMAIN%' ),
+	'parent_item_colon' => esc_html__( 'Parent Videos Category:', '%TEXTDOMAIN%' ),
+	'edit_item' => esc_html__( 'Edit Videos Category', '%TEXTDOMAIN%' ),
+	'update_item' => esc_html__( 'Update Videos Category', '%TEXTDOMAIN%' ),
+	'add_new_item' => esc_html__( 'Add New Videos Category', '%TEXTDOMAIN%' ),
+	'new_item_name' => esc_html__( 'New Videos Category', '%TEXTDOMAIN%' ),
 	'separate_items_with_commas' => esc_html__( 'Separate videos categories with commas', '%TEXTDOMAIN%' ),
 	'add_or_remove_items' => esc_html__( 'Add or remove videos categories', '%TEXTDOMAIN%' ),
 	'choose_from_most_used' => esc_html__( 'Choose from the most used videos categories', '%TEXTDOMAIN%' ),
+	'not_found' => esc_html__( 'No categories found', '%TEXTDOMAIN%' ),
 	'menu_name' => esc_html__( 'Categories', '%TEXTDOMAIN%' ),
 );
 
@@ -57,6 +58,7 @@ $labels = array(
 	'separate_items_with_commas' => esc_html__( 'Separate tags with commas', '%TEXTDOMAIN%' ),
 	'add_or_remove_items' => esc_html__( 'Add or remove tags', '%TEXTDOMAIN%' ),
 	'choose_from_most_used' => esc_html__( 'Choose from the most used tags', '%TEXTDOMAIN%' ),
+	'not_found' => esc_html__( 'No tags found', '%TEXTDOMAIN%' ),
 	'menu_name' => esc_html__( 'Tags', '%TEXTDOMAIN%' ),
 );